requirements.txt
requests>=2.28.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
newspaper3k>=0.2.8
readability-lxml>=0.8.1
feedparser>=6.0.10
openai>=1.0.0
edge-tts>=6.1.0
pyyaml>=6.0
click>=8.1.0
tenacity>=8.2.0
python-dateutil>=2.8.0
